Conflicts:
	src/wwwroot/.eslintignore
	src/wwwroot/app/directives/datepickerwrapper/DatePickerWrapperDirective.js
	src/wwwroot/app/index.html
	src/wwwroot/app/main.js
	src/wwwroot/app/views/candidate/candidate.controller.js
	src/wwwroot/app/views/vacancy/vacancy.view.html
	src/wwwroot/bower.json
	src/wwwroot/config/config.webpack/webpack.global.js

+++ src/wwwroot/app/services/VacancyService.js
@@ -12,13 +12,13 @@ export default class VacancyService {
    }
 
    getVacancy(vacancyId) {
-      var additionalUrl = VACANCY_URL + vacancyId;
+      const additionalUrl = VACANCY_URL + vacancyId;
       return _HttpService.get(additionalUrl);
    }
 
    saveVacancy(entity) {
-      if (entity.Id !== undefined) {
-         var additionalUrl = VACANCY_URL + entity.Id;
+      if (entity.id) {
+         const additionalUrl = VACANCY_URL + entity.id;
          return _HttpService.put(additionalUrl, entity);
       } else {
          return _HttpService.post(VACANCY_URL, entity);
@@ -26,7 +26,11 @@ export default class VacancyService {
    }
 
    deleteVacancy(entity) {
-      var additionalUrl = VACANCY_URL + entity.Id;
-      _HttpService.remove(additionalUrl, entity);
+      if (entity.id === undefined) {
+         throw new Error('Id should be specified');
+      } else {
+         const additionalUrl = VACANCY_URL + entity.id;
+         _HttpService.remove(additionalUrl, entity);
+      }
    }
 }
